requirements.txt
browser-use
python-docx
